Move Cosmos query components to queries:: module
