+++ crates/cosmos/cosmos-client-components/src/components/client.rs
@@ -65,16 +65,16 @@ use crate::impls::connection::connection_handshake_payload::BuildCosmosConnectio
 use crate::impls::connection::init_connection_options::ProvideCosmosInitConnectionOptionsType;
 use crate::impls::packet_fields::CosmosPacketFieldReader;
 use crate::impls::packet_from_ack::BuildCosmosPacketFromWriteAck;
-use crate::impls::query_block::QueryCometBlock;
-use crate::impls::query_chain_id::QueryChainIdWithChainHandle;
-use crate::impls::query_chain_status::QueryChainStatusWithChainHandle;
-use crate::impls::query_consensus_state_height::QueryConsensusStateHeightFromChainHandle;
-use crate::impls::query_packet_commitments::QueryCosmosPacketCommitments;
-use crate::impls::query_received_packet::QueryReceivedPacketWithChainHandle;
-use crate::impls::query_send_packet::QueryCosmosSendPacket;
-use crate::impls::query_send_packets::QuerySendPacketsConcurrently;
-use crate::impls::query_unreceived_packet::QueryUnreceivedCosmosPacketSequences;
-use crate::impls::query_write_ack_event::QueryWriteAckEventFromChainHandle;
+use crate::impls::queries::block::QueryCometBlock;
+use crate::impls::queries::chain_id::QueryChainIdWithChainHandle;
+use crate::impls::queries::chain_status::QueryChainStatusWithChainHandle;
+use crate::impls::queries::consensus_state_height::QueryConsensusStateHeightFromChainHandle;
+use crate::impls::queries::packet_commitments::QueryCosmosPacketCommitments;
+use crate::impls::queries::received_packet::QueryReceivedPacketWithChainHandle;
+use crate::impls::queries::send_packet::QueryCosmosSendPacket;
+use crate::impls::queries::send_packets::QuerySendPacketsConcurrently;
+use crate::impls::queries::unreceived_packet::QueryUnreceivedCosmosPacketSequences;
+use crate::impls::queries::write_ack_event::QueryWriteAckEventFromChainHandle;
 use crate::impls::receive_packet_message::BuildCosmosReceivePacketMessage;
 use crate::impls::receive_packet_payload::BuildCosmosReceivePacketPayload;
 use crate::impls::send_messages_as_tx::SendMessagesToTxContext;
