+++ crates/cosmos/cosmos-client-components/src/components/ibc_client.rs
@@ -6,8 +6,8 @@ use hermes_relayer_components::chain::traits::queries::consensus_state::Consensu
 
 use crate::impls::client::create_client_message::BuildCosmosCreateClientMessage;
 use crate::impls::connection::connection_handshake_message::BuildCosmosConnectionHandshakeMessage;
-use crate::impls::query_client_state::QueryCosmosClientStateFromChainHandle;
-use crate::impls::query_consensus_state::QueryCosmosConsensusStateFromChainHandle;
+use crate::impls::queries::client_state::QueryCosmosClientStateFromChainHandle;
+use crate::impls::queries::consensus_state::QueryCosmosConsensusStateFromChainHandle;
 
 pub struct CosmosIbcClientComponents;
