+++ crates/cosmos/cosmos-client-components/src/impls/mod.rs
@@ -5,18 +5,7 @@ pub mod client;
 pub mod connection;
 pub mod packet_fields;
 pub mod packet_from_ack;
-pub mod query_block;
-pub mod query_chain_id;
-pub mod query_chain_status;
-pub mod query_client_state;
-pub mod query_consensus_state;
-pub mod query_consensus_state_height;
-pub mod query_packet_commitments;
-pub mod query_received_packet;
-pub mod query_send_packet;
-pub mod query_send_packets;
-pub mod query_unreceived_packet;
-pub mod query_write_ack_event;
+pub mod queries;
 pub mod receive_packet_message;
 pub mod receive_packet_payload;
 pub mod send_messages_as_tx;
