+++ crates/cosmos/cosmos-client-components/src/impls/queries/write_ack_event.rs
